FINAL, comments added, SOLUTION updated, and that's it

+++ ssh_keys_rotation.sh
@@ -6,6 +6,8 @@ if [ $# -ne 1 ]; then
   exit 1
 fi
 
+PRIVATE_IP=$1
+
 # checks if the old key exists
 if [ -e private_key.pem ]; then
   mv private_key.pem old_key.pem
@@ -14,7 +16,17 @@ fi
 
 # generate new key-pair
 ssh-keygen -t rsa -b 2048 -f private_key.pem -N ""
+if [ $? -ne 0 ]; then
+  echo "new key-pair generation failed"
+  exit 1
+fi
+
 # copy the public key to authorized_keys in private ec2
-cat private_key.pem.pub | ssh -o StrictHostKeyChecking=accept-new -i old_key.pem ubuntu@$1 "cat > ~/.ssh/authorized_keys"
+cat private_key.pem.pub | ssh -o StrictHostKeyChecking=accept-new -i old_key.pem ubuntu@"$PRIVATE_IP" "cat > ~/.ssh/authorized_keys"
+if [ $? -ne 0 ]; then
+  echo "failed to copy private key to private instance authorized_keys"
+  exit 1
+fi
+
 # remove old key
 rm -f old_key.pem old_key.pem.pub
